+++ packages/eyes-storybook/test/it/getStories.it.test.js
@@ -98,10 +98,6 @@ describe('getStories', () => {
             res.parameters = {
               fileName: './test/fixtures/appWithStorybook/index.js',
               framework: 'react',
-              options: {
-                hierarchyRootSeparator: '|',
-                hierarchySeparator: {},
-              },
               ...parameters,
             };
           } else {