+++ packages/dom-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+## 4.7.6
+
+- update @applitools/eyes-common@3.18.5
+
 ## 4.7.5
 
 - update @applitools/eyes-common@3.18.4