+++ packages/eyes-selenium-jest-perf/CHANGELOG.md
@@ -1,7 +1,6 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## Unreleased
 
 ## [4.5.98](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-selenium-jest-perf@4.5.97...@applitools/eyes-selenium-jest-perf@4.5.98) (2019-12-02)