+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+## 2.12.2
+- branching base commit support
+- make functions in runners 
+
 ## 2.12.1
 - Fixed regression which broke xpath selectors when looking up the location of an element
 
@@ -15,7 +21,7 @@
 ## 2.11.7
 - Fixed error when getting images on Safari. Related to [Trello 99](https://trello.com/c/C0oZf2oc)
 
-## [unversioned] - 
+## [un versioned] - 
 ### Fixed
 - Crop safari url bar and status bar on mobile devices (IOS). [Trello 939](https://trello.com/c/SM80YzdM)